refactor(core): Modularize application components
Restructure core application logic into distinct, single-responsibility
classes: KeystrokeListener, FocusTracker, and SnippetHandler. This
enhances modularity, testability, and maintainability of the codebase.

The Application class now orchestrates these components, delegating
specific tasks such as keystroke monitoring, active window focus
tracking, and snippet replacement logic to the respective handlers.
This separation of concerns clarifies the role of each module and
simplifies future development and debugging.

Signal/slot connections have been updated to reflect the new component
interactions. Import statements across the affected modules were
adjusted to maintain correct package-relative paths.

Also added test architecture for unittesting later on
